pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hjepa-moe"
version = "0.1.0"
description = "Hierarchical JEPA with Mixture-of-Experts Predictors"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "einops>=0.7",
]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "wandb"]

[tool.setuptools.packages.find]
include = ["hjepa_moe*"]
